requirements.txt
APScheduler==3.6.3
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
idna==3.3
ipykernel==6.10.0
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
parso==0.8.3
peewee==3.14.10
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
validators==0.18.2
wcwidth==0.2.5
wget==3.2
